requirements.txt
transformers==4.29.0
openai
huggingface_hub>=0.14.1
diffusers 
accelerate 
datasets 
torch
soundfile 
sentencepiece 
opencv-python
python-dotenv
protobuf==3.20.0
streamlit